+++ extensions/mozilla-firefox/src/constants.ts
@@ -6,15 +6,18 @@ export const SEARCH_ENGINE: { [key: string]: string } = {
   duckduckgo: `https://duckduckgo.com/?q=`,
 };
 
-export const DownloadText = `
+const makeDownloadText = (installer: string, installerUrl: string) => `
   # 🚨Error: Mozilla Firefox browser is not installed
   ## This extension depends on Mozilla Firefox browser. You must install it to continue.
   
-  If you have [Homebrew](https://brew.sh/) installed then press ⏎ (Enter Key) to install Mozilla Firefox browser.
+  If you have [${installer}](${installerUrl}) installed then press ⏎ (Enter Key) to install Mozilla Firefox browser.
   [Click here](https://www.mozilla.org/en-US/firefox/new/) if you want to download manually.
   
   [![Mozilla Firefox](https://mozilla.design/files/2019/10/logo-firefox.svg)]()
 `;
+
+export const DownloadText = makeDownloadText("Homebrew", "https://brew.sh/");
+export const DownloadTextWindows = makeDownloadText("Winget", "https://aka.ms/winget");
 
 export const NoBookmarksText = `
 # 🚨Error: Mozilla Firefox browser has no bookmarks. Please add some bookmarks to continue using this command.